% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.to.seasonal}
\alias{aggregate.to.seasonal}
\title{Aggregate 8am to 8am data to yearly sums and seasonal sums}
\usage{
\method{aggregate}{to.seasonal}(aggregate88)
}
\arguments{
\item{aggregate88}{The 8-8 aggregated data to be further aggregated to seasonal
#@example aggregated.seasonal <- aggregate.to.seasonal(data.tables$`DeBilt_550_N_8-8daily_precip`)}
}
\description{
Since the variability is seasonal dependent we would like to be able to compare certain seasons only.
}
\details{
Calculate seasonal sums

When data availability is below 80% in a season (i.e. >18 days missing), or a year (i.e. > 73 days missing), the aggregation becomes -9999.
}
\author{
Jurian
}
