% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateOperations.R
\name{aggregate288}
\alias{aggregate288}
\title{Aggregate hourly data to daily from 8am to 8am the next day}
\usage{
aggregate288(data.container)
}
\arguments{
\item{data.container}{The R object containing all timeseries and metadata}
}
\description{
Rain is measured hourly in the automatic weather stations, this needs to be converted to sums from 8 am to 8 am on the next day.
}
\details{
Daily aggeration
}
\author{
Lotte, Jurian & Hidde
}
