% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateOperations.R
\name{aggregate2Seasonal}
\alias{aggregate2Seasonal}
\title{Aggregate 8 am to 8 am data to seasonal sums}
\usage{
aggregate2Seasonal(data.container)
}
\arguments{
\item{data.container}{The R object containing all timeseries and metadata}
}
\description{
Since the variability is seasonal dependent we would like to be able to compare certain seasons only. The function takes the R object, calculates aggregates and returns this in the object, included updated meta data.
}
\details{
When data availability is below 80% in a season (i.e. >18 days missing), or a year (i.e. > 73 days missing), the aggregation becomes NA.
}
\author{
Jurian and Lotte
}
