% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateOperations.R
\name{aggregate2Year}
\alias{aggregate2Year}
\title{Aggregate 8 am to 8 am data to yearly sums}
\usage{
aggregate2Year(data.container)
}
\arguments{
\item{data.container}{The R object containing all timeseries and metadata}
}
\description{
The function takes the R object, calculates aggregates and returns this in the object, included updated meta data.
}
\details{
When data availability is below 80% in a season (i.e. >18 days missing), or a year (i.e. > 73 days missing), the aggregation becomes NA.
}
\author{
Jurian and Lotte
}
