% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateOperations.R
\name{aggregateTo.88}
\alias{aggregateTo.88}
\title{Aggregate hourly data to daily from 8am to 8am the next day}
\usage{
aggregateTo.88(obj, all.stations = TRUE, sta_type = "AWS", var_id = "RH",
  sta_id = NULL)
}
\arguments{
\item{obj}{The R object containing all timeseries and metadata}

\item{all.stations}{default is "TRUE" means all stations are aggregated. If all.stations = FALSE, an array of sta_ID needs to be provided.}

\item{sta_type}{default is "AWS", but could be extended to for instance "WOW" in the future.}

\item{var_id}{default is "RH" for hourly rainfall.}

\item{sta_id}{defines the string of sta_id's that need to be aggregated. Only applies when all.stations = FALSE. 
#@example aggregated88 <- aggregate.to.88(obj=obj)}
}
\description{
Rain is measured hourly in the automatic weather stations, this needs to be converted to sums from 8 am to 8 am on the next day.
}
\details{
Daily aggeration
}
\author{
Lotte, Jurian & Hidde
}
