% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateOperations.R
\name{aggregateTo.seasonal}
\alias{aggregateTo.seasonal}
\title{Aggregate 8am to 8am data to yearly sums and seasonal sums}
\usage{
aggregateTo.seasonal(obj, all.stations = TRUE, sta_type = "AWS",
  var_id = "RD", sta_id = NULL)
}
\arguments{
\item{obj}{The R object containing all timeseries and metadata}

\item{all.stations}{default is "TRUE" means all stations are aggregated. If all.stations = FALSE, an array of sta_ID needs to be provided.}

\item{sta_type}{default is "AWS", but could be extended to for instance "WOW" in the future.}

\item{var_id}{default is "RD" for daily rainfall.}

\item{sta_id}{defines the string of sta_id's that need to be aggregated. Only applies when all.stations = FALSE. 
#@example aggregated.seasonal <- aggregate.to.seasonal(obj)}
}
\description{
Since the variability is seasonal dependent we would like to be able to compare certain seasons only. The function takes the R object, calculates aggregates and returns this in the object, included updated meta data.
}
\details{
Calculate seasonal sums

When data availability is below 80% in a season (i.e. >18 days missing), or a year (i.e. > 73 days missing), the aggregation becomes -9999.
}
\author{
Jurian and Lotte
}
