% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averagingOperations.R
\name{average.spatial}
\alias{average.spatial}
\title{Create an average over several stations}
\usage{
average.spatial(timeseries)
}
\arguments{
\item{timeseries}{A list of data.tables, each a timeseries of structure <datetime, value>}
}
\value{
A single timeserie of structure <datetime, value>
}
\author{
Hidde and Jurian
}
