% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_break_detection_buishand.R
\name{break.detection}
\alias{break.detection}
\title{Break detection with Buishand method}
\usage{
break.detection(df.difference)
}
\arguments{
\item{df.difference}{An input table with two columns: year and difference}
}
\description{
Function for break detection in a yearly or seasonal series that uses the Buishand cumulative test. It needs a table with two columns (year and value). the input is temporarily downloaded by the dummy data
}
\details{
Break detection
}
\author{
Antonello Squintu
}
