% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{buishand}
\alias{buishand}
\title{Break detection with Buishand method}
\usage{
buishand(id, input)
}
\arguments{
\item{id}{id}

\item{input}{input}
}
\description{
Function for break detection in a yearly or seasonal series that uses the Buishand cumulative test. It needs a table with two columns (datetime and value). the input is temporarily downloaded by the dummy data
}
\author{
Antonello Squintu
}
