% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{db.close}
\alias{db.close}
\title{Close the connection to the MySQL database}
\usage{
db.close(db)
}
\arguments{
\item{db}{Data base}
}
\value{
TRUE, invisibly
}
\description{
This closes the connection, discards all pending work, and frees resources (e.g., memory, sockets).
}
\author{
Jurian and Hidde
}
