% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{db.execute}
\alias{db.execute}
\title{Execute a database operation}
\usage{
db.execute(FUN, ...)
}
\arguments{
\item{FUN}{One of the db.<operation> functions}

\item{...}{Arguments to pass to the function}
}
\value{
Output of the function
}
\description{
This function is meant for quickly doing a database operation without needing to open and close a connection explicitly. This is all done automatically.
}
\author{
Jurian
#@example data.container <- db.execute(db.select.all, "1day", "H", "RD")
#@example data.id <- db.execute(db.new.data.id)
}
