% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{db.insert.update.timeseries}
\alias{db.insert.update.timeseries}
\title{Insert or update a specific timeseries based on its meta-data}
\usage{
db.insert.update.timeseries(db, meta, timeseries)
}
\arguments{
\item{db}{Handle to MySQL database, taken from db.setup()}

\item{meta}{An object of type mqm.meta.timeseries}

\item{timeseries}{An object of type mqm.data.timeseries, data.table of structure <datetime, value>}
}
\seealso{
db.setup()
}
\author{
Jurian
}
