% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{db.new.data.id}
\alias{db.new.data.id}
\title{Fetch a new data ID from the database}
\usage{
db.new.data.id(db)
}
\arguments{
\item{db}{Handle to MySQL database, taken from db.setup()}
}
\value{
A new and unique data ID
}
\description{
a function
}
