% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{db.select.timeseries}
\alias{db.select.timeseries}
\title{Query the database for timeseries data and metadata}
\usage{
db.select.timeseries(db, station.IDs, time.interval, type, element)
}
\arguments{
\item{db}{Handle to MySQL database, taken from db.setup()}

\item{station.IDs}{A vector of unique station ID's (called "codes" in the DB)}

\item{time.interval}{One of {"1hour", "1day", "month", "season, "year"}}

\item{type}{One of {"N", "H"} (case insensitive)}

\item{element}{One of {"RH", "RD", "RR"} (case insensitive)}
}
\value{
An object of type "mqm.data.container" which contains a list of timeseries and metadata on those series.
#@example data.container <- db.select.timeseries(db, c(260, 324, 343, 340), "1hour", "H", "RH")
}
\description{
a function
}
\seealso{
db.setup()
}
\author{
Jurian
}
