% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{db.setup}
\alias{db.setup}
\title{Setup a connection to the MySQL database}
\usage{
db.setup()
}
\value{
An object of class "MySQLConnection" and "RMySQL"
}
\description{
This function returns a handle which can be used later to query the db.
}
\author{
Jurian and Hidde
}
