% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacontainerOperations.R
\name{merge.mqm.data.container}
\alias{merge.mqm.data.container}
\title{Merge two data containers}
\usage{
merge.mqm.data.container(a, b)
}
\arguments{
\item{a}{Fist data container}

\item{b}{Second data container}
}
\value{
A new data container containing the meta data from both containers and their respective time interval data
#@example 1hour.and.1day.data <- merge.mqm.data.container(1hour.data, 1day.data)
}
\description{
Note that behavior for merging two idential time intervals is not well defined! I.e. 1day with 1day or 1hour with 1hour, etc...
}
\author{
Jurian
}
