% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_relativedifference.R
\name{relative.difference}
\alias{relative.difference}
\title{Relative difference}
\usage{
relative.difference(AWS_timeserie, Man_timeserie)
}
\arguments{
\item{AWS_timeserie}{The input contains 1 timecolumn and only 1 value column with AWS precipitation averages named 'average', and NA values are indicated with '-9999'.}

\item{Man_timeserie}{The input contains 1 timecolumn and only 1 value column with Man precipitation averages named 'average', and NA values are indicated with '-9999'.}
}
\description{
#@example input; files <- grab.test.data(); AWS_timeserie <- files$Almelo_664_N_seasonal_precip[,c(1,3)]; Man_timeserie <- files$Enschede_665_N_seasonal_precip[,c(1,3)]; names(AWS_timeserie) <- names(Man_timeserie) <- c("time", "average")
}
