% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{station.info}
\alias{station.info}
\title{Get station information from database}
\usage{
station.info(db)
}
\arguments{
\item{db}{Handle to MySQL database, taken from db.setup()}
}
\description{
get the metadata for all stations
}
\author{
Marieke
}
