% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesOperations.R
\name{timeseries.intersection}
\alias{timeseries.intersection}
\title{Intersect multiple timeseries on their common datetime value}
\usage{
timeseries.intersection(timeseries)
}
\arguments{
\item{timeseries}{A list of data.tables, each a timeseries of structure <datetime, value>}
}
\value{
A list of data.tables, each a timeseries of structure <datetime, value>. Only with observations that all timeseries have in common
}
\description{
Supposed to be very fast and efficient, i.e. very fast even on millions of observations
}
\details{
NB. This function assumes that all timeseries are complete!
}
\author{
Jurian and Hidde
}
