% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesOperations.R
\name{timeseries.relative.difference}
\alias{timeseries.relative.difference}
\title{Relative difference of timeseries}
\usage{
timeseries.relative.difference(timeserie1, timeserie2)
}
\arguments{
\item{timeserie1}{timeserie as data.table with first column datetime and second column (average) value. Possible input is AWS_timeserie.}

\item{timeserie2}{timeserie as data.table with first column datetime and second column (average) value. Possible input is MAN_timeserie.}
}
\description{
Calculate the relative difference between two timeseries as 100% * (timeseries2 - timeseries1)/timeseries1.
}
\details{
the input should have certain column names and identical timeseries. If not, the function stops with an error that identifies the problem.
}
\author{
Lotte and Jurian
}
