% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.to.88}
\alias{aggregate.to.88}
\title{Aggregate hourly data to daily from 8am to 8am the next day}
\usage{
\method{aggregate}{to.88}(hourly)
}
\arguments{
\item{hourly}{The hourly data to be aggregated to 8-8 data}
}
\description{
Rain is measured hourly in the automatic weather stations, this needs to be converted to sums from 8 am to 8 am on the next day.
}
\details{
Daily aggeration
}
\author{
Lotte
}
