% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakDetection.R
\name{break.detection}
\alias{break.detection}
\title{Break detection with Buishand method}
\usage{
break.detection(series1, name, type, radius = 2, plot.score = FALSE)
}
\arguments{
\item{series1}{time series to check, it must be a table with two columns (datetime and value)}

\item{name}{name (or code) of the station}

\item{type}{temporal aggregation of the input {'y', 'djf', 'mam', 'jja', 'son'}}

\item{radius}{radius (in years) of the running window that is used for the identification of local minima/maxima}

\item{plot.score}{if TRUE it produces a plot for the visualization of the breaks}
}
\description{
Function for break detection in a yearly or seasonal series that uses the Buishand cumulative test. It needs a table with two columns (datetime and value). the input is temporarily downloaded by the dummy data
}
\author{
Antonello and Hidde
}
