% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakDetection.R
\name{break.detection}
\alias{break.detection}
\title{Break detection with Buishand method}
\usage{
break.detection(series1, range = 2, plot.score = FALSE)
}
\description{
Function for break detection in a yearly or seasonal series that uses the Buishand cumulative test. It needs a table with two columns (datetime and value). the input is temporarily downloaded by the dummy data
}
\author{
Antonello Squintu and Hidde
}
