% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{break.info}
\alias{break.info}
\title{Get break detection information from database}
\usage{
break.info(comb_name)
}
\arguments{
\item{comb_name}{combine name}
}
\description{
get the break detection information per combine name
}
\author{
Else
}
