% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{db.select.all}
\alias{db.select.all}
\title{Query the database for hourly}
\usage{
db.select.all(db, time.period, station.type, element.name)
}
\arguments{
\item{db}{Handle to MySQL database, taken from db.setup()}

\item{time.period}{One of {"hour", "day", "month", "year"} NB. Season not supported ATM!}

\item{station.type}{One of {"validated", "derived"}}

\item{element.name}{Name of type of data; e.g. "rh" for hourly rainfall or "t" for temperature}
}
\seealso{
db.setup()
}
\author{
Jurian and Hidde
}
