% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{db.select.all}
\alias{db.select.all}
\title{Query the database for hourly}
\usage{
db.select.all(db, time.interval, type, element)
}
\arguments{
\item{db}{Handle to MySQL database, taken from db.setup()}

\item{time.interval}{One of {"1hour", "1day", "month", "season, "year"}}

\item{type}{One of {"N" (Manual), "H" (Automatic)} Case insensitive}

\item{element}{One of {"RH" (Precipitation, originated from hourly data), "RD" (Precipitation, from daily data), "RR" (Precipitation, from radar data)} Case insensitive
#@example data.container <- db.execute(db.select.all, "1hour", "H", "RH")}
}
\description{
a function
}
\seealso{
db.setup()
}
\author{
Jurian and Hidde
}
