% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{db.select.all}
\alias{db.select.all}
\title{Query the database for hourly}
\usage{
db.select.all(db, time.interval, type, element)
}
\arguments{
\item{db}{Handle to MySQL database, taken from db.setup()}

\item{time.interval}{One of {"1hour", "1day", "month", "season, "year"}}

\item{type}{One of {"N", "H"} Case insensitive}

\item{element}{One of {"RH", "RD", "RR"} Case insensitive}
}
\description{
a function
}
\seealso{
db.setup()
}
\author{
Jurian and Hidde
}
