% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{db.select.timeseries}
\alias{db.select.timeseries}
\title{Query the database for timeseries data and metadata}
\usage{
db.select.timeseries(db, station.IDs, type.ID, element.ID, series.type)
}
\arguments{
\item{db}{Handle to MySQL database, taken from db.setup()}

\item{series.type}{One of {"validated", "derived", "series"}}

\item{stationIDs}{A vector of unique station ID's (called "codes" in the DB)}

\item{typeID}{Unique identifier for type (e.g. 1 for "N", 2 for "H")}

\item{elementID}{Unique identifier for element (e.g. 1 for "RH", 2 for "RD")}
}
\value{
An object of type "mqm.data.container" which contains a list of timeseries and metadata on those series.
}
\seealso{
db.setup()
}
\author{
Jurian
}
