% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{grab.test.data}
\alias{grab.test.data}
\title{Compute seasonal data}
\usage{
grab.test.data()
}
\arguments{
\item{aggregate88}{The 8-8 aggregated data to be further aggregated to seasonal}

\item{hourly}{The hourly data to be aggregated to 8-8 data
#@example aggregated88 <- aggregate.to.88(data.tables$`DeBilt_260_H_hourly_precip`)}
}
\description{
Load in test data from Mariekes directory
}
\details{
#@import data.table
}
