% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{station.info}
\alias{station.info}
\title{Get station information from database}
\usage{
station.info()
}
\value{
TRUE, invisibly
}
\description{
get the metadata for all stations
}
\author{
Marieke
}
