% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseOperations.R
\name{station.nearby}
\alias{station.nearby}
\title{Get information from nearby stations from the database}
\usage{
station.nearby(db, code_real)
}
\arguments{
\item{code_real}{code like 260_H}
}
\description{
get the metadata for all stations, input looks like code_real="260_H"
}
\author{
Marieke
}
